+++ apps/mis-web/src/apis/api.mock.ts
@@ -111,6 +111,7 @@ export const mockApi: MockApi<typeof api> = {
     return { items: ["test.test", "test1.test2"]};
   },
 
+
   getAllTenants: async () => (
     {
       totalCount: 3,