+++ apps/mis-web/src/pages/api/admin/changePassword.ts
@@ -10,45 +10,48 @@
  * See the Mulan PSL v2 for more details.
  */
 
-import { route } from "@ddadaal/next-typed-api-routes-runtime";
-import { changePassword as libChangePassword } from "@scow/lib-auth"; 
+import { typeboxRoute, typeboxRouteSchema } from "@ddadaal/next-typed-api-routes-runtime";
+import { changePassword as libChangePassword } from "@scow/lib-auth";
+import { Type } from "@sinclair/typebox";
 import { authenticate } from "src/auth/server";
 import { PlatformRole } from "src/models/User";
 import { publicConfig, runtimeConfig } from "src/utils/config";
 
 // 此API用于账户管理员修改其他任意用户的密码。
 // 没有权限返回undefined
-export interface ChangePasswordAsPlatformAdminSchema {
+export const ChangePasswordAsPlatformAdminSchema = typeboxRouteSchema({
 
-  method: "PATCH";
+  method: "PATCH",
 
-  body: {
-    identityId: string;
-    oldPassword: string;
+  body: Type.Object({
+    identityId: Type.String(),
+    oldPassword: Type.String(),
     /**
      * @pattern ^(?=.*\d)(?=.*[a-zA-Z])(?=.*[`~!@#\$%^&*()_+\-[\];',./{}|:"<>?]).{8,}$
      */
-    newPassword: string;
-  };
+    newPassword: Type.String({
+      pattern:  "^(?=.*\\d)(?=.*[a-zA-Z])(?=.*[`~!@#\\$%^&*()_+\\-[\\];',./{}|:\"<>?]).{8,}$",
+    }),
+  }),
 
   responses: {
     /** 更改成功 */
-    204: null;
+    204: Type.Null(),
 
     /** 用户未找到 */
-    404: null;
+    404: Type.Null(),
 
     /** 密码不正确 */
-    412: null;
-    
+    412: Type.Null(),
+
     /** 本功能在当前配置下不可用。 */
-    501: null;
-  }
-}
+    501: Type.Null(),
+  },
+});
 
 
-export default /* #__PURE__*/route<ChangePasswordAsPlatformAdminSchema>(
-  "ChangePasswordAsPlatformAdminSchema", async (req, res) => {
+export default /* #__PURE__*/typeboxRoute(
+  ChangePasswordAsPlatformAdminSchema, async (req, res) => {
 
     if (!publicConfig.ENABLE_CHANGE_PASSWORD) {
       return { 501: null };