+++ apps/mis-web/src/pages/api/admin/updateBlockStatus.ts
@@ -10,23 +10,24 @@
  * See the Mulan PSL v2 for more details.
  */
 
-import { route } from "@ddadaal/next-typed-api-routes-runtime";
+import { typeboxRoute, typeboxRouteSchema } from "@ddadaal/next-typed-api-routes-runtime";
 import { asyncClientCall } from "@ddadaal/tsgrpc-client";
 import { AdminServiceClient } from "@scow/protos/build/server/admin";
+import { Type } from "@sinclair/typebox";
 import { authenticate } from "src/auth/server";
 import { PlatformRole } from "src/models/User";
 import { getClient } from "src/utils/client";
 
-export interface UpdateBlockStatusSchema {
-  method: "PUT";
+export const UpdateBlockStatusSchema = typeboxRouteSchema({
+  method: "PUT",
 
   responses: {
-    200: null;
-  }
-}
+    200: Type.Null(),
+  },
+});
 const auth = authenticate((info) => info.platformRoles.includes(PlatformRole.PLATFORM_ADMIN));
 
-export default route<UpdateBlockStatusSchema>("UpdateBlockStatusSchema",
+export default typeboxRoute(UpdateBlockStatusSchema,
   async (req, res) => {
 
     const info = await auth(req, res);